Moved code to the right files
